Added an echo server and client based on UNIX domain sockets.
The code now includes an echo server and client that use UNIX sockets
for local process communication. To accommodate this change, we
refactored the previous echo server and client to separate the code
needed for Internet sockets from the protocol logic and processing.
Both the Internet and UNIX echo code use the same underlying socket
code for the echo protocol, but have specialized code to create and
close these different types of sockets.